Move config package to service/internal/config
